requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-multiselectfield==0.1.12
djangorestframework==3.13.1
flake8==5.0.4
mccabe==0.7.0
Pillow==9.2.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyPDF2==2.9.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.3.0
tzdata==2022.1
